search history from recents screen. Fix history being reinserted instead of updating existing chapter history.

+++ core/common/src/main/java/io/silv/common/model/ChapterResource.kt
@@ -6,6 +6,6 @@ interface ChapterResource {
     val title: String
     val mangaId: String
     val volume: Int
-    val chapter: Long
+    val chapter: Double
     val url: String
 }
